OPENJPA-213
git-svn-id: https://svn.apache.org/repos/asf/openjpa/trunk@658197 13f79535-47bb-0310-9956-ffa450edef68

+++ openjpa-jdbc/src/main/java/org/apache/openjpa/jdbc/sql/DBDictionary.java
@@ -1499,7 +1499,7 @@ public class DBDictionary
             case JavaTypes.BIGDECIMAL:
                 if (storeLargeNumbersAsStrings)
                     return getPreferredType(Types.VARCHAR);
-                return getPreferredType(Types.DOUBLE);
+                return getPreferredType(Types.NUMERIC);
             case JavaTypes.NUMBER:
                 if (storeLargeNumbersAsStrings)
                     return getPreferredType(Types.VARCHAR);
